requirements.txt
Django==1.11
django-rest-framework
